Implemented password protection- unfortunately this does not work on wwwx due to old version of PHP. Fixed minor bugs having to do with recording.

+++ serverSide/RestInterface.php
@@ -14,32 +14,35 @@ $pathInfo = explode('/', $_SERVER['PATH_INFO']);
 // See if we're manipulating items
 if ($pathInfo[1] == "Items"){ // first entry is blank (starts with '/')
    
+   // Make sure we are allowed to edit items
+   $listSlug = $pathInfo[2];
+   $list = Database::getList($listSlug);
+   $listData = $list->getData();
+   verifyThatListIsAccessible($listData); // exits if credentials invalid
+
    // Item creation and update
    if ($_SERVER['REQUEST_METHOD'] == "POST"){
       
-      $listSlug = $pathInfo[2];
       $listOrder = $pathInfo[3];
       
       // Are we creating an item?
       if ($listOrder == "new"){
          
          // We are updating a list item
-         $list = Database::getList($listSlug);
          $item = $list->insertItem(trim($_REQUEST['content'])); // since we don't support whitespace, might as well trim
          exit();
       
       // Are we deleting an item?
       } else if ($listOrder == "delete"){
          
-         $listSlug = $pathInfo[2];
          $listOrder = $pathInfo[4];
-         $list = Database::getList($listSlug);
+         
          $item = $list->getItem($listOrder);
          $item->delete();
       
       // Otherwise, we are updating an item
       } else {
-         $list = Database::getList($listSlug);
+         
          $item = $list->getItem($listOrder);
          
          if (array_key_exists('content', $_REQUEST)){
@@ -70,14 +73,34 @@ if ($pathInfo[1] == "Items"){ // first entry is blank (starts with '/')
       
    // Otherwise, we're updating the list data
    } else {
+      
+      // Only allow this if the list is already unlocked
       $list = Database::getList($pathInfo[2]);
-      $list->rename($_REQUEST['title']);
+      $listData = $list->getData();      
+      verifyThatListIsAccessible($listData); // exits if credentials invalid
+      
+      // If we're requesting to remove password protection
+      if (array_key_exists('passwordProtected', $_REQUEST) && $_REQUEST['passwordProtected'] == 0){
+         $list->unlock();
+         exit();
+         
+      // If we're requesting to set a password
+      } else if (array_key_exists('newPassword', $_REQUEST)){
+         $hash = password_hash($_REQUEST['newPassword'], PASSWORD_DEFAULT);
+         print($hash);
+         $list->lock($hash);
+         
+      // Otherwise, we're updating the title.
+      } else if (array_key_exists('title', $_REQUEST)){
+         $list->rename($_REQUEST['title']);
+      }
       exit();
-   }
+   }//end if not new list
 
    
-// GET to RestInterface/listSlug gives back a JSON with the list name and
-// the list items.
+// GET to RestInterface/listSlug is the main access point for getting list information.
+// It acts differently depending on whether the list is password protected, and on
+// whether a password was provided.
 } else if ($_SERVER['REQUEST_METHOD'] == "GET"){
 
    // Check that the REST identifier (the "slug") is valid. The path starts with '/',
@@ -103,33 +126,101 @@ if ($pathInfo[1] == "Items"){ // first entry is blank (starts with '/')
       exit();
    }
    
-   $listItems = $list->getItemsData();
-   
-   // Create an object to be serialized and sent back with the name of the list and
-   // an array containing the item data.
-   $toSend = new FullListResponse($listData->title, $listItems);
+   // Check whether list is not password protected
+   if ($listData->passwordProtected == 0){      
+      $listItems = $list->getItemsData();
       
-   // Create the JSON response
-   header("Content-type: application/json");
-   print(json_encode($toSend));
+      // Create an object to be serialized and sent back with the name of the list and
+      // an array containing the item data.
+      $toSend = new FullListResponse($listData->title, $listItems, $listData->passwordProtected);
+         
+      // Create the JSON response
+      header("Content-type: application/json");
+      print(json_encode($toSend));
+      exit();
+   
+   // If the list is password protected
+   } else {
+      // If a password was not provided, this is a request to find out if a password
+      // is needed.
+      if (!array_key_exists('password', $_REQUEST)){
+         
+         // Create the JSON response
+         header("Content-type: application/json");
+         print(json_encode(new PasswordProtectedResponse()));
+         exit();
+         
+      // If a password was provided, and the list is protected...
+      } else {
+         
+         // Compare to the actual password
+         if (password_verify($_REQUEST['password'], $listData->hashedPassword)){
+         
+            // Get the data and send it back
+            $listItems = $list->getItemsData();
+            $toSend = new FullListResponse($listData->title, $listItems, $listData->passwordProtected);
+               
+            // Create the JSON response
+            header("Content-type: application/json");
+            print(json_encode($toSend));
+            exit();            
+            
+         } else {
+            http_response_code(401);
+            print('Incorrect password.');
+            exit();
+         }
+      }//end if password provided
+   }//end if password protected   
 }//end GET
+
+/**
+   Utility function that sends back an error response if the list is locked and the
+   password does not match.
+*/
+function verifyThatListIsAccessible($listData){
+   
+   // If list is not locked, or if the password is correct.
+   if ($listData->passwordProtected == 0 
+   || (array_key_exists('password', $_REQUEST) && password_verify($_REQUEST['password'], $listData->hashedPassword))){
+      return;
+      
+   // Otherwise, give an error response
+   } else {
+      print($listData->hashedPassword);
+      print("\n");
+      print($_REQUEST['password']);
+      http_response_code(401);
+      print('Invalid credentials.');
+      exit();
+   }
+}
 
 // Our wwwx server does not support anonymous classes, hence we define the classes
 // we need for responses below.
 class FullListResponse {
-      public $name;
-      public $items;
-      
-      function __construct($listName, $listItems){
-         $this->name = $listName;
-         $this->items = $listItems;
-      }
+   public $name;
+   public $items;
+   public $passwordProtected;
+   
+   function __construct($listName, $listItems, $passwordProtected = 0){
+      $this->name = $listName;
+      $this->items = $listItems;
+      $this->passwordProtected = $passwordProtected;
+   }
 }
 class ListSlugResponse {
    public $slug;
    
    function __construct($slug){
          $this->slug = $slug;
+   }
+}
+class PasswordProtectedResponse {
+   public $passwordProtected;
+   
+   function __construct(){
+      $this->passwordProtected = 1;
    }
 }
 ?>
